Moved spg_database to using Spacegroup class instead of struct.
This change allowed for simpler pybind exposure of the Spacegroup
information of a Lattice to Python.

Changed the eigenvector cost calculation in the Munkres' permutation
calculation, as the full eigenvectors for each mode are orthogonal
not necessarily the, e.g., N individual ion 3-element eigenvectors
for the phonons in an N-ion system.