Fixed major issue with pybind not enforcing c-ordered arrays
The pybind numpy array wrapper py::array_t<T> allows for an
additional template parameter which is intended to specify the
element ordering of the numpy array. I had used py::array::c_style
under the assumption that it would convert Fortran-style (MATLAB)
numpy arrays automatically for me; it did not.
Adding the py::array:forcecast parameter did not cause pybind to
instruct numpy to change from F_CONTIGUOUS to C_CONTIGUOUS.

The fix to this issue is to further support Python buffer objects.
Since pybind can supply `itemsize`, `ndim`, `shape`, and `stride`
along with the pointer we can work out how to transform the passed
numpy array ourselves.
There is now a constructor for ArrayVector and LDVec/LQVec objects
which takes the pointer, shape, and stride and copies the numpy
array into a contiguous row-ordered object for interal processing.

This fix has the added benefit of allowing for non-contiguous
data or Q arrays to be passed from Python, and saves a copy which
would have otherwise been used to make a contiguous version before
passing on to the C++ methods.