Consensus sorting is implemented but does not work.
Whether the inability of consensus sorting to produce reasonable
output is a result of bugs or intrinsic to the idea is not yet
clear. More work will be done to sort this out.

As a new feature, grids with lower dimensionality than their
container can now be interpolated (for 3D at least). This is
achieved by forcing otherwise out-of-bounds interpolation points
to have an exact match for the singleton grid dimensions.
This should be useful, e.g., when a 1-D or 2-D system is to be
compared to measured data as the instrumental effects can still
be modeled properly without also needing to replicate the
system response up to 3-D.