Starting towards Irreducible Brillouin zone grids.
One solution to the sorting issue might be reducing the space that needs 
to be sorted as much as possible by moving to the Irreducible Brillouin 
zone. This requires a scheme to identify a wedge of reciprocal space 
that can be replicated via the pointgroup symmetries -- a prototype of 
which is included in the new class IrreducibleBrillouinZone which should 
probably be merged with BrillouinZone if/when the scheme works.

The pointgroup code from spglib has been cleaned up a bit and expanded 
as necessary. My Symmetry and PointSymmetry classes have been moved from 
raw array to vector<array<>> internal storage.