Irreducible reciprocal space wedge introduced.
The BrillouinZone object now has the ability to calculate its own
irreducible wedge of reciprocal space and to use this information
to produce the Irreducible Brillouin Zone.
The BrillouinZoneGrid3 object leverages this to drastically reduce
(in high symmetry cases) the number of points in the mapped grid,
and can interpolate scalar, vector, and matrix objects on the
irreducible grid including proper handling of the pointgroup
operation required to move each Q into the irreducible Brillouin
zone -- Q = Rq + tau, where tau is a reciprocal lattice vector,
q is a point in the irreducible Brillouin zone, and R is one of
the pointgroup operations.

As currently implemented, the grid still occupies the full
Brillouin zone (with only the irreducible part mapped to gridded
data) and therefore dense grids can take considerable memory and
computation time (for determining which parts are in the
irreducible part of the zone).
A future improvement will move to grids constructed from the
vertices of the irreducible Brillouin zone by tetrahedron
subdivision. This should improve the memory use and setup time and
might help with edge effects (as all grid points can be inside or
on the surface of the IR-Bz).