New object BrillouinZoneMesh with attendant Python object BZMesh*
Now that the irreducible BrillouinZone is identifiable, it may be
important to restrict the interpolation fixed-points to lie on or
within it, in order to avoid issues with interpolating across the
irreducible Brillouin zone boundary.
Using a tetrahedron mesh is one way to ensure that the set of
interpolation fixed-points conforms to the IrBz polyhedron.

Generation and refinement of such a mesh is now handled by a
slightly modified version of TetGen. Finding the containing
tetrahedron of a point within the IrBz is a non-trivial task and,
for the time being, CGAL is being used to accomplish this.

Functions exist to interpolate at points within the meshed IrBz,
though they have not been extensively tested yet.