Moved BrillouinZone objects to holding two Polyghedron objects
Since the first Brillouin zone and irreducible Brillouin zone are
both polyhedra, it made sense to treat both the same way and
create a new Polyhedron class.
Now creating the vertex-to-facet and facet-to-vertex indexing
arrays is done for both the first and irreducible Brillouin zones,
including per-facet vertex ordering, automatically in the C++ code.
This enabled rather significant clean-up of the Python plotting
routine in addition to the simplification of the BrillouinZone
header and source files.