Bug fixes for low-symmetry pointgroup irreducible wedge searches.
For spacegroups with pointgroup symmetry having any combination of
1 and -1 operations and up to a single unique 2-fold operation the
previous irreducible reciprocal space wedge search failed to set
anything for the irreducible Brillouin zone polyhedron.
In the case of only 1 and -1 the irreducible Brillouin zone *is*
the first Brillouin zone, and the updated methods set this
correctly.
For 2+-fold pointgroups the irreducible Brillouin zone occupies
less reciprocal space volume than the first Brillouion zone and
the updated methods work even for a single unique 2-fold axis.

The polyhedron class is improved to calculate its own volume, this
improves the BrillouinZone method which checks for non-zero-
volume irreducible Brillouin zones when determining which normal
vectors can describe the irreducible reciprocal space bounding
planes. The volume method is exported via pybind as well.

Debugging output has been improved to use macro functions, so that
the calls do not get compiled if a DEBUG flag is not set. This
enables easier-than-line-comment-toggling switching from verbose
output to detected exception output.
The exception handling of ArrayVector errors is slightly improved
with the goal of simplifying bugs encountered via Python where it
is nontrivial to use a C debugger.