BrillouinZone fixes for inversion-free systems
The irreducible Brillouin zone is constructed under the assumption
that a system has time-reversal and/or space-inversion symmetry as
in this case it forms a convex polyhedron and identifying
relationships between vertices and planes is simplified.

For systems lacking time-reversal and space-inversion symmetry
such a constructed polyhedron is only half of the irreducible
Brillouin zone, with the other half being the mirror of the convex
polyhedron.
For such systems the BrillouinZone object now returns the correct
non-convex irreducible Brillouin zone polyhedron, which enables,
e.g., creation of tetrahedral meshes which account for the lack of
inversion symmetry.