Remove dependency on CGAL. Fixes for MATLAB & Python interfaces
Relying on CGAL was becoming increasingly less useful as, after
triangulating with TetGen, we would have had to construct a
CGAL Triangulation structure by hand. As CGAL uses an infinite
vertex to describe the surface triangles on our polyhedra and
TetGen does not, this was a complicated mess.

CGAL triangulations are gone and the TetGen triangulation result
gets wrapped up in a much nicer to deal with C++ object.