Changes for a sub-optimal fix to the ir reciprocal space search
The search now has an option to prefer the primitive basis vectors
as in-perpendicular-plane vectors for each pointgroup symmetry
operation. And another option to allow parallel v_i and u_j which
was previously disallowed in all cases.
The suboptimal part of the fix is that some spacegroups have their
irreducible Brillouin zone found with one set of options while
others require a different set, so the constructor tries each
combination until an appropriately-sized "irreducible" polyhedron
is found (one with a volume equal to the full first Brillouin
zone volume divided by the number of pointgroup symmetry
operations). If no such ir-Bz is found an error is thrown.