Progress towards brute-force irBZ determination.
Polyhedra can now return the points halfway along each of their
edges. Along with the vertices and face centres, these are all
of the possible vertices (along with Gamma) of any irreducible
Brillouin zone.
We can verify the previous statement by checking that every point
is stationary for at least one pointgroup operation other than 1
or -1.

Since we know that the irBz volume must be the first Bz volume
divided by the number of pointgroup symmetry operations, we should
be able to find an appropriately sized polyhedron and then verify
that it tiles the first Bz.
Hopefully there are strategies to reduce the complexity of the
brute force method -- for the FCC lattice there are 74 total points
on face centres, vertices, and edge centres; a polyhedron needs
at least four points but could use all of the points, so there are
{74:4} + {74:5} + ... {74:74} combinations of points that we
might need to check, where {N:M} is the binomial coefficient.
That is 1150626 + 16108764 + ... + 1 = 1.89E22 (!).
The real number to check can be reduced if we ensure that we
never have three colinear points in the polhedron, but that check
might be equally as bad computationally.