Separated out CMakeLists into directory-specific files
The goal of which is to simplify project maintenance and code
portability.

Work on different first Brillouin zone construction method, taking
inspiration from voro++. Rather than finding the vertices of all
orders of Brillouin zones and then pulling out just the first zone
vertices, construct a probably-too-large rectangular-prism and
then use Brillouin zone planes to cut it down to size.
For this to work we either need to use another library (voro++)
which can lead to interface issues, or define plane-slicing of a
polyhedron.
The second approach is now implemented and needs to be tested.