New first Brillouin zone construction method works
Though remains not-yet-fully tested. Since it does not look for
intersections of all possible half-tau planes it is significantly
faster than the old algorithm.

Work needs to shift back to determining the irreducible Brillouin
zone. The brute-force method can find a correct-volume polyhedron
but not necessarily one which fills the first Brillouin zone after
application of all symmetry elements.