New Polyhedron methods to search for intersections
Looking for the intersection between two convex polyhedra is
unfortunately rather complicated. The implemented method seems to
work in some cases but fails in others.

A bug by-which the face normals were opposite in sign from what
they should be for all of the irreducible-space wedge normals of
the irreducible Brillouin zone polyhedron has been fixed.
This enables the volume of the polyhedron to be calculated
correctly if it is translated (necessary for the intersection test)
and also enables correct determination of, e.g., the polyhedron
centroid.

Improvements to the debugging output macros and printer which
will help clean-up code using them, especially by reducing the
need to use debug_exec to build-up output strings.