Fixed the intersection test for polyhedra.
With the fix the Brillouinzone object's method to test if the
current irreducible Brillouin zone is correct seems to work.

New python test to check whether the irreducible Brillouin zone
determination works for all 530 Hall number spacegroups.
As implemented, the correct irBZ is found for slightly less than
half of all spacegroups -- and the brute-force search does not
improve the situation as it has a bug causing it to hang for
low-symmetry pointgroups.

Starting work on implementing an exact calculation of the important
high-symmetry points, which are the vertices of the irBZ. This
method is going to be terribly ugly and impossible to maintain.