New pseudo brute force irreducible wedge search method.
The new method is not yet bug free but is considerably more
promising than the older brute force method or the explicit method
(which sadly won't work, as we need more than just one of each
distinct high symmetry point in the first Brillouin zone).
There is almost certainly a bug in new Polyhedron code intended to
eliminate the possibility of zero-area polygon faces. Previously
a check was made whether a face had at least 3 points, now an
algorithm attempts to determine if any number of points happen to
be colinear or non-coplanar. The other possible bug location is
in determining the polygon vertex ordering for each face where I
removed a face-vector normalization while looking for a bug caused
by a "face" comprised of three colinear points equally spaced
along their line, such that the centre point had a face-vector of
zero and a normalized vector of (NaN, NaN, NaN).