Improvements to deal with floating point comparison issues
The new brute force irBz method is still not bug free. While
investigating its issue a more insidious issue became apparent.
The issue is a result of how floating point numbers are compared
for equivalence at the moment -- using a scaled machine precision
that does not account for the number of operations contributing
to the error in each number.
A real fix for this issue might be difficult but the temporary
solution is to accept a 10-fold larger difference between floating
point numbers before claiming they are different.