Reciprocal lattice vector transformation bug fixed.
For a given transformation matrix P which converts the direct
space lattice vectors A into their primitive equivalent basis
vectors A' via A' = A P, the components of a direct space vector
expressed in units of the conventional cell basis vectors can be
transformed to units of the primitive cell basis vectors by
x' = inverse(P) x. The transformation methods were performing this
transform correctly.
Where there was an error was in transforming reciprocal space
vectors. If the basis vectors of the conventional reciprocal
lattice form the columns of B then the primitive reciprocal lattice
vectors form the columns of B' = B tranpose(inverse(P))
and the components of a vector expressed in the basis vectors of
the conventional reciprocal lattice can be transformed to units
of the primitive reciprocal lattice by q' = tranpose(P) q. The
error was in omitting the transpose. For most centred lattices
this makes no difference as their transform matrices, P, are
symmetric. For the Hexagonal setting of the rhombohedral lattice,
however, this is not the case.

Improvements to the irreducible Brillouin zone method have reduced
the number of lattices for which it fails. Still more work is
needed for some Orthorhombic lattices (122, 209, 212, 334-336) and
most Rhombohedral lattices (433, 435-437, 439, 441, 443-446 ...)
and *probably* all monoclinic lattices.