Improvements to irreducible wedge search. 13/530 still failing.
The irreducible Brillouin zone found for some spacegroups, e.g.,
R-3m, are suboptimal since they cross high-symmetry-point
connecting-lines. There are alternative irreducible volumes which
do not cross between Brillouin zone faces unnecessarily, but they
may not be (or are not in the case of R-3m) a convex polyhedron.
A test for this condition could be whether the vertices of an
irreducible Brillouin zone face contain a non-contiguous subset
of the vertices of a full Brillouin zone face.

If a non-convex irreducible Brillouin zone could be found by some
algorithm (or supplied by a user) we would need some way to
determine whether a given point is inside of the polyhedron.
The simplest solution might be to identify the convex hull of the
polyhedron and the convex polyhedra which form the difference
between the convex hull and irreducible polyhedron.
In any case, this requires a more-robust set of computational
geometry routines, which may require siginificant effort.