Fixed bug in src/bz_wedge_3.cpp which caused too few wedge normals
The code at the end of wedge_brute_force to deal with the case of
ignored space inversion was written with the intent of adding an
additional wedge normal to divide the irreducible polyhedron.
The appropriate normal to add was found, and the polyhedron was
correctly subdivided but due to an oversight the new normal was
added to an empty ArrayVector. This one-vector ArrayVector was
then set as the sole wedge normal for the BrillouinZone.
This had the effect of preventing ir_moveinto from working as
anticipated, which caused (ir_)interpolate_at to fail due to
points persisting outside of the irreducible polyhedron.

The fix was to implement a new method which returns the negated
normals of the irreducible polyhedron which are not full Brillouin
zone normals, instead of relying on the stored values.
At the moment we still store the wedge normals, but it might be
worthwhile to only ever find them from the polyhedron.