Fixed undefined symbol error in python module.
Importing the python module raised an ImportError:

[path]/[module].so: undefined symbol _ZN17PolyhedronTrellisIdEC2ERK10Polyhedrond

The mangled symbol name gives a hint to the source of the problem
but demangling it with `c++filt` was necessary to track down my
absent-minded mistake. The demangled symbol is

PolyhedronTrellis<double>::PolyhedronTrellis(Polyhedron const&, double)

which is the intitialiser of the PolyhedronTrellis class object.
I had written the class to be definately typed and then added-in
the ability for it to hold an ArrayVector<T> of data, which added
the template parameter to the whole class. The absent-minded
mistake was in leaving the now-templated initialiser definition in
a source file instead of a header file.

Moving the template code into a header file has solved the issue.