Fixed PolyhedronTrellis vertex indexing
The code to generate all trellis intersection vertices and to
compute their linear indices disagreed on the way to store the
three-dimensional array in a vector.
They must use the same convention as the PolyhedronTrellis nodes,
which at the moment are using column-major (Fortran style) instead
of row-major ordering.

Since, after object creation, we only ever compute a node index
from its subscripts (found by 'binning' a point) it probably
doesn't matter whether we stick with column- or row-major order
here. But if someone decides to change it in the future then all
PolyhedronTrellis vertex and node related code must change
together.