The nested tetrahedra object fixed
When creating the tetrahedra tree structure all branched vertices
were being stored correctly but the branches themselves were not.
As a result only the top-level nodes were saved and the finest-
grained leaves were lost.
By changing the construction slightly the branches are fully
subdivided before being added to the root level and everything
seems to be working as expected.