Found and bypassed cause of null trellis node access
In some cases determining the truncated node polyhedron creates
an open polyhedron (with at least one face missing). This then
causes tetgen to fail to produce *any* triangulation and a null
node was stored in order to deal with the (maybe) possible case
of a truncated node *really* resulting in an empty polyhedron.

The bypass implemented checks for an empty tetgen triangulation
and tries creating a convex hull polyhedron from the truncated
node vertices before creating a new triangulation from that.

To truely fix the issue one needs to determine *why* the
truncation of the node cube fails to produce a closed polyhedron.