Fixed an error in a previous Polyhedron bugfix.
The polyhedron bisect static method as originally written was
inadvertantly appending duplicate intersection points to the
vertex list. That is, no check was performed to verify that the
found vertex was actually new before pushing it onto the back of
the polyhedron vertices list. The multiple copies of a vertex
would then cause future algorithms to fail as, e.g., two faces
with a shared edge might not be found if one (or both) end points
used different indexed copies of the same point in the vertices
list.

To fix the original problem I added a check that the found vertex
should only be retained if it did not already exist in the list.
The error was that the way the check was added the pre-existing
vertex was also omitted from the newly-created face, which caused
its own problems for irreducible Brillouin zone creation for
some spacegroups.

The bisect method now correctly avoids duplicating polyhedron
vertices *and* adds the correct index to the face vertex list.