Verify minimu cmake version. Simplify BrillouinZone python tests
In trying to get Github workflow continuous integration working a few issues with CMakeLists.txt were identified. Namely that it may be necessary to specify how cmake handles relative paths when encountering `target_sources` directives. This also re-verifies that 3.13.0 is the minimum required version of cmake for this to build as prior to that release all paths were treated as absolute.

Some aspect of printing the failing BrillouinZone spacegroups to console is not safe for non-unicode supporting terminals -- likely the character Å -- as a quick fix the python test no longer outputs any 'known failure' information.

Presently the Github server fails to run catch2 via ctest where a numerical error is thrown. Sorting this issue will likely require more effort but will be necessary if the workflow based continuous integration will ever work.