First steps in rebasing the Polyhedron class
A small bug causes an error in info_update, debug_update, verbose_update if the
terminal emulator reports its width as 0. Now such terminals have the largest
possible width for formatted output.

The most important method of the Polyhedron class is `bisect`, which was written
under the assumption that all polyhedra would always be convex hulls.
One shortcut was to avoid fully tracking the vertices associated with each facet
when they were cut and instead rely upon the convex hull initializer to sort
things out.
This approach fails if polyhedron is/was not convex -- as is often the case for
mirrored 'half'-irreducible polyhedra. Such a cut polyhedron could leave one or
more dangling vertices after cutting which would lead to mistakes.
Fixing this issue requires that we keep better track of which vertices actually
contribute to each face, which the Polyhedron::bisect method is now better able
to do.
If we keep track of the face vertices correctly we no longer need to store or
track the plane points or normals (as they can always be calculated from the
list of face vertices). This refactoring will remove them from the class
properties which should improve the legibility of the methods.

Some terminal emulators do not set their width/height?!

Polyhedron::bisect keeps better track of vertices_per_facet

Polyhedron now tracks vertices_per_facet correctly for new test cases