Improvements to irBZ finding
The `wedge_brute_force` algorithm mistakenly treated 2-fold
rotations and mirror operatiors in the same manner.
A list of prescribed wedge-normals was used in either case, however
those normals only make sense for 2-fold rotations.
The algorithm has been updated to not apply the method of predetermined
wedge normals for mirror planes but instead to use the mirror plane
normals (their 'stationary' axes) as their contributing wedge normals.

This change improves the number of test lattices for which finding an
irreducible Brillouin zone works. There are still a handful of lattices
in the test suite which fail, at least one of which due to how close to
tetragonal it is despite being monoclinic with unique angle 90.498. In
this case the vertices of the irreducible Brillouin zone are found but
the convex hull Polyhedron algorithm fails, misidentifying very small
faces next to a large nearly-square face as a second nearly-square
face behidn the first.
Fixing this might be as simple as ensuring Polyhedra only have unique
faces in their output, but might require keeping lattice information
with the polyhedron or modification of the wedge search algorithm to
identify retainted-points with faces of the full Brillouin zone.