Fixed issue caused by using pymatgen to find lattice Hall symbols
Pymatgen wraps spglib to determine the symmetry of a given lattice.
By a default pymatgen or spglib setting the determined symmetry is
standardised. This is probably a sensible default behaviour but in the
case of a structure with only a subset of atoms (e.g., just the magnetic
ones) spglib may pick a higher-symmetry spacegroup for its output.
One of the AFLOW structures is monoclinic with only a small deviation
from orthorhombic; in at least this case spglib found an 'equivalent'
orthorhombic lattice and I mistakenly combined the monoclinic lattice
parameters with the orthorhombic spacegroup the irreducible Brillouin
zone creation test. Naturally this failed.

Since the AFLOW lattice symmetries are known there is no good reason to
use pymatgen/spglib to parse the data and determine the symmetries. The
JSON lattices list has been regenerated using the symmetry information
directly from the AFLOW files.
The offending lattice is now identified as monoclinic and we can
construct a correct irreducible Brillouin zone for it.