Fixed remaining issue with ir BZ creation for some cubic systems
When determining which high-symmetry points remain in the cut-down
polyhedron checks are performed to ensure that one or more normals
does not wipe-out all points because its pointing the wrong way.
In the mirror-plane case the check was performed but the recorded
normal direction was not modified, so that cutting the full Brillouin
zone polyhedron resulted in a null polyhedron.

With the fix in place all test lattices are working.