Move default pyton build to Release.
Add optimisation flags to Release target. Change the format of the build
information which is recorded in the compiled python module slightly.

Remove all tetgen mesh refinement as it was causing significant trellis
generation time slowdowns for quartz. If the non-Delaunay-ness of the
generated mesh is a problem we likely have a bigger problem
interpolating between the corners of cubes anyway.