Fix errors in determining irreducible q.
The determination of irreducible q from Q and a set of pointgroup
operations, R in G, was incorrectly treating the matrix multiplication
of R and q. Since the pointgroup operations are defined in real space it
is necessary to apply their transpose to the reciprocal vectors, which
was not done. This error is only apparent for spacegroups with
non-orthogonal basis vectors where transpose(R) != R for all R in G.

An additional error preventing the correct determination of an
irreducible q was in checking whether a given Q is within the allowed
irreducible recriprocal space wedge. Spacegroups without inversion
symmetry do not have equivalent properties at Q and -Q; this fact does
not also imply that the pointgroup operations of such a spacegroup can
not replicate the positive irreducible 'cone' to all reciprocal space --
which was an incorrect assumption within the in-wedge determining code.
Where a point was deemed 'in wedge' if Q or -Q were calculated inside
for inversion-free spacegroups, now that is only true if necessary for
the irreducible wedge to fill the first Brillouin zone.