Static MacOS python OpenMP linkage. Move from 4 to 3 elements
The interface from MacOS MATLAB 2019b through python to brille
does not work unless the static version of the OpenMP library
is linked into _brille.

While interpolating any complex-valued objects should be allowed
to have an overall arbitrary phase difference. Along with this
change it no longer makes sense to treat eigenvectors and 'normal'
vectors differently, so the interpolation_data class does not.
This means it only needs to know about scalar, vector, and matrix
elements in the stored array and it enforces that there are 3N
vector elements and 9M matrix elements, which is a significant
change from before.
This will necessitate a new version number to indicate the API
change, but it can wait until all interpolation objects use the
new approach.