Removed extraneous CMake flag specifications. Remove C++17 features.
Setting extra flags 'by hand' inside of the CMakeLists file was not the
right way to do things. It prevented the 'correct' flags from being set
in VisualStudio, for example.
Removing the flag specifications and relying instead on the CMake build
type targets revealed a few problems in the polyhedron bisection code
which were resolved by switching to std::remove_if instead of erasing
std::vector elements by their iterators.

The PrimitiveTranspose class contained `constexpr static const 
std::array`
protected properties to represent the transformation matrices between
primitive lattices and their conventional unit cells.
This was a purely C++17 feature which some compilers (gcc, MSVS, intel)
were happy to include when the C++14 standard was specified.
The matrices are now accessible only through get_P and get_invP methods
of initalised PrimitiveTransform objects, but this is the only way they
were accessed previously. Hopefully this will resolve runtime issues 
with the python module compiled with some gcc and intel compilers, and
also compilation issues with some versions of clang.

The PolyhedronTrellis class had a very inefficient bit in its 
initialiser
which has now been resolved. There is still a rather inefficient part
used to relate vertices of the cubic node polyhedra back to the full 
list
of all retained vertices. It should be possible to improve this as well,
but it will require a bit of thought.