Faster PolyhedronTrellis creation time
When a trellis node extends beyond the bounding polyhedron and has been
cut, it will still contain at least one vertex which is a trellis
intersection point. In order to keep only unique vertices for all nodes
it is necessary to check first *if* a triangulated vertex is one of the
trellis intersection points and then determine *which* one it is.

The check was performed quickly before by checking against only the 8
vertices of the to-be-cut node. If this found a single match a
comparison was performed against *all* kept trellis intersection points
to find the right index -- since the number of trellis intersections
scales with its density more accurate trellises required a more
expensive comparison.

The newly improved method keeps track of the mapped trellis intersection
point indices which surround a given node. It uses this list of mapped
indices to check whether a triangulated vertex is one of the kept
trellis intersection points -- this check will always be against less
than 8 points, otherwise the node would not have been cut and
triangulated. Directly finding which mapped index the vertex matches
eliminates the need for the second density-dependent check, making the
new method significantly faster than the old one.