Correct BrillouinZoneTrellis3 hidden tests
These tests were missed when interpolate_at was renamed ir_interpolate 
at and the new non-irreducible interpolate_at function was added for the 
BrillouinZoneTrellis3 class.
Attempting to use the new interpolate_at with a irreducible
Brillouin zone limited trellis will not work for any pointgroup with 
non-identity members, so the tests failed.