Fix double free error in TetGen
When adding the Gamma point to a triangulation TetGen uses some 
additional internal arrays. These arrays are allocated with new and 
deallocated with delete before and after their use in a number of 
functions. When the call to tetrahedralize is finished the tetgen object 
is deleted calling its freememory() method. Within that method both 
arrays are deleted again and the checks to prevent this (that they're 
not NULL) fail.
This is either a bug in TetGen or in how brille calls tetrahedralize 
when an extra point is required. The current solution of removing the 
delete statements from freememory fixes the double free issue but might 
lead to memory leaks in untested cases.

There is now a working test to ensure that the Gamma point is one of the 
vertices of a previously-problematic Trellis.