Improved BrillouinZone::moveinto
The old moveinto method could fail to find an equivalent reduced
q for some spacegroups and Q directions.
Specifically, points along the line from (000) to (111) in a
hexagonal spacegroup pass the first Brillouin zone boundary when
they reach (1/3,1/3,1/3). Points between (1/3,1/3,1/3) and
(1/2,1/2,1/2) were correctly moved into the first Brillouin zone
by the old algorithm, e.g., a single translation of (-1,0,0) *or*
(0,-1,0) places those points within the 1stBz.
Points between (1/2,1/2,1/2) and (1,1,1) were not moved into the
1stBz correctly due, at least in part, to the need to translate
by both (001) and either of (100) or (010).

The new moveinto method finds the closest-integer (hkl) to Q and
uses this as an initial guess for tau. For primitive cubic,
tetragonal, and orthorhombic systems this *is* tau and the
algorithm is done. For all systems a check is made whether the
resulting q = Q - tau is outside of the 1stBz and, if so, it
performs the same steps as the old method.
For the case of a hexagonal lattice, points between (1/2,1/2,1/2)
and (1,1,1) are first mapped to (-1/2,-1/2,-1/2) to (0,0,0);
points beyond (2/3,2/3,2/3) are immediately mapped to the 1stBz
while points between (1/2,1/2,1/2) and (2/3,2/3,2/3) need further
processing. (1/2+x,1/2+x,1/2+x)->(-1/2+x,-1/2+x,-1/2+x) which is
outside of the 1stBz but a translation by (100) or (010) brings it
within.

There could still be cases where the new moveinto method will fail
but as it passes all tests at the moment, this will hopefully not
be the case.