Finish updating grid types to hold InterpolationData objects
All interpolation objects now have a property which is of the same
InterpolationData class. This should simplify any future modifications
to how the interpolation data is handled.
At present those objects which can determine sorting permuations of
their held data *should* still be able to do so, but moving the sorting
from the interpolation objects to the InterpolationData class might
make sense in the future.

Since the InterpolationData object requires only a three-element
specifier, all grid-like objects need the same input which is an API
modification. The version has been incremented accordingly.