Compilation (#10)
* Fixed for compilation errors under GCC 9.2

* Static MacOS python OpenMP linkage. Move from 4 to 3 elements

The interface from MacOS MATLAB 2019b through python to brille
does not work unless the static version of the OpenMP library
is linked into _brille.

While interpolating any complex-valued objects should be allowed
to have an overall arbitrary phase difference. Along with this
change it no longer makes sense to treat eigenvectors and 'normal'
vectors differently, so the interpolation_data class does not.
This means it only needs to know about scalar, vector, and matrix
elements in the stored array and it enforces that there are 3N
vector elements and 9M matrix elements, which is a significant
change from before.
This will necessitate a new version number to indicate the API
change, but it can wait until all interpolation objects use the
new approach.

* Removed extraneous CMake flag specifications. Remove C++17 features.

Setting extra flags 'by hand' inside of the CMakeLists file was not the
right way to do things. It prevented the 'correct' flags from being set
in VisualStudio, for example.
Removing the flag specifications and relying instead on the CMake build
type targets revealed a few problems in the polyhedron bisection code
which were resolved by switching to std::remove_if instead of erasing
std::vector elements by their iterators.

The PrimitiveTranspose class contained `constexpr static const 
std::array`
protected properties to represent the transformation matrices between
primitive lattices and their conventional unit cells.
This was a purely C++17 feature which some compilers (gcc, MSVS, intel)
were happy to include when the C++14 standard was specified.
The matrices are now accessible only through get_P and get_invP methods
of initalised PrimitiveTransform objects, but this is the only way they
were accessed previously. Hopefully this will resolve runtime issues 
with the python module compiled with some gcc and intel compilers, and
also compilation issues with some versions of clang.

The PolyhedronTrellis class had a very inefficient bit in its 
initialiser
which has now been resolved. There is still a rather inefficient part
used to relate vertices of the cubic node polyhedra back to the full 
list
of all retained vertices. It should be possible to improve this as well,
but it will require a bit of thought.

* Faster PolyhedronTrellis creation time

When a trellis node extends beyond the bounding polyhedron and has been
cut, it will still contain at least one vertex which is a trellis
intersection point. In order to keep only unique vertices for all nodes
it is necessary to check first *if* a triangulated vertex is one of the
trellis intersection points and then determine *which* one it is.

The check was performed quickly before by checking against only the 8
vertices of the to-be-cut node. If this found a single match a
comparison was performed against *all* kept trellis intersection points
to find the right index -- since the number of trellis intersections
scales with its density more accurate trellises required a more
expensive comparison.

The newly improved method keeps track of the mapped trellis intersection
point indices which surround a given node. It uses this list of mapped
indices to check whether a triangulated vertex is one of the kept
trellis intersection points -- this check will always be against less
than 8 points, otherwise the node would not have been cut and
triangulated. Directly finding which mapped index the vertex matches
eliminates the need for the second density-dependent check, making the
new method significantly faster than the old one.

* Correct BrillouinZoneTrellis3 hidden tests

These tests were missed when interpolate_at was renamed ir_interpolate 
at and the new non-irreducible interpolate_at function was added for the 
BrillouinZoneTrellis3 class.
Attempting to use the new interpolate_at with a irreducible
Brillouin zone limited trellis will not work for any pointgroup with 
non-identity members, so the tests failed.

* Add Gamma point to all Trellises and  Fix double free error in TetGen

When adding the Gamma point to a triangulation TetGen uses some 
additional internal arrays. These arrays are allocated with new and 
deallocated with delete before and after their use in a number of 
functions. When the call to tetrahedralize is finished the tetgen object 
is deleted calling its freememory() method. Within that method both 
arrays are deleted again and the checks to prevent this (that they're 
not NULL) fail.
This is a bug in TetGen. The current solution of removing the 
delete statements from freememory fixes the double free issue and has
been endorsed by Hang Si via email.

There is now a working test to ensure that the Gamma point is one of the 
vertices of a previously-problematic Trellis.

* Improved BrillouinZone::moveinto

The old moveinto method could fail to find an equivalent reduced
q for some spacegroups and Q directions.
Specifically, points along the line from (000) to (111) in a
hexagonal spacegroup pass the first Brillouin zone boundary when
they reach (1/3,1/3,1/3). Points between (1/3,1/3,1/3) and
(1/2,1/2,1/2) were correctly moved into the first Brillouin zone
by the old algorithm, e.g., a single translation of (-1,0,0) *or*
(0,-1,0) places those points within the 1stBz.
Points between (1/2,1/2,1/2) and (1,1,1) were not moved into the
1stBz correctly due, at least in part, to the need to translate
by both (001) and either of (100) or (010).

The new moveinto method finds the closest-integer (hkl) to Q and
uses this as an initial guess for tau. For primitive cubic,
tetragonal, and orthorhombic systems this *is* tau and the
algorithm is done. For all systems a check is made whether the
resulting q = Q - tau is outside of the 1stBz and, if so, it
performs the same steps as the old method.
For the case of a hexagonal lattice, points between (1/2,1/2,1/2)
and (1,1,1) are first mapped to (-1/2,-1/2,-1/2) to (0,0,0);
points beyond (2/3,2/3,2/3) are immediately mapped to the 1stBz
while points between (1/2,1/2,1/2) and (2/3,2/3,2/3) need further
processing. (1/2+x,1/2+x,1/2+x)->(-1/2+x,-1/2+x,-1/2+x) which is
outside of the 1stBz but a translation by (100) or (010) brings it
within.

There could still be cases where the new moveinto method will fail
but as it passes all tests at the moment, this will hopefully not
be the case.

* Finish updating grid types to hold InterpolationData objects

All interpolation objects now have a property which is of the same
InterpolationData class. This should simplify any future modifications
to how the interpolation data is handled.
At present those objects which can determine sorting permuations of
their held data *should* still be able to do so, but moving the sorting
from the interpolation objects to the InterpolationData class might
make sense in the future.

Since the InterpolationData object requires only a three-element
specifier, all grid-like objects need the same input which is an API
modification. The version has been incremented accordingly.